// src/app/page/master/map-paket-medical-to-produk/map-paket-medical-to-produk.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapPaketMedicalToProduk {
	
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
 }